refactor(audio-bridge): removed make_plugin usage
