+++ jarust_plugins/examples/audio_bridge_list.rs
@@ -1,6 +1,6 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
-use jarust_plugins::audio_bridge::AudioBridge;
+use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
 use tracing_subscriber::EnvFilter;
 
 #[tokio::main(flavor = "current_thread")]
