+++ jarust_plugins/src/audio_bridge/mod.rs
@@ -1,34 +1,5 @@
 pub mod events;
 pub mod handle;
+pub mod jahandle_ext;
 pub mod messages;
 pub mod responses;
-
-use self::events::AudioBridgePluginData;
-use self::events::AudioBridgePluginEvent;
-use self::handle::*;
-use jarust::japrotocol::EstablishmentProtocol;
-use jarust::japrotocol::JaEventProtocol;
-use jarust::japrotocol::JaResponseProtocol;
-use jarust::prelude::*;
-use jarust_make_plugin::make_plugin;
-
-make_plugin!(AudioBridge, "janus.plugin.audiobridge");
-
-impl AudioBridge for JaSession {
-    type Event = (AudioBridgePluginEvent, Option<EstablishmentProtocol>);
-    type Handle = AudioBridgeHandle;
-
-    fn parse_audio_bridge_message(message: JaResponse) -> JaResult<Self::Event> {
-        let msg = match message.janus {
-            JaResponseProtocol::Event(JaEventProtocol::Event { plugin_data }) => (
-                serde_json::from_value::<AudioBridgePluginData>(plugin_data.data)?.event,
-                message.establishment_protocol,
-            ),
-            _ => {
-                tracing::error!("unexpected response");
-                return Err(JaError::UnexpectedResponse);
-            }
-        };
-        Ok(msg)
-    }
-}
